Experiment: Auto-convert mixed CRLF+LF to only CRLF
Maybe this check+conversion is not needed -- just remember that if there
are ^M characters at EOL, then the file has mixed EOLs. Try to determine
when they occur and fix the problem at its root if possible. If not
possible, maybe this patch will help prevent it in the future.

In short: Use this patch if no other solution exists. Avoid modifying
the buffer unless it is absolutely necessary.

See:
http://vim.wikia.com/wiki/File_format
http://vim.wikia.com/wiki/Automatically_reload_files_with_mixed_line-endings_in_DOS_fileformat